refactor(provider): remove unused ListRunner utilities and tests
- Removed `ListRunner` generic utilities and associated test cases to streamline the codebase.
- Improves provider maintainability by eliminating unused artifacts.
